backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
astroid==2.14.2
async-timeout==4.0.2
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Deprecated==1.2.13
dill==0.3.6
Django==4.1.7
django-celery-results==2.4.0
django-cors-headers==3.14.0
django-extensions==3.2.1
django-redis==5.2.0
djangorestframework==3.14.0
isort==5.12.0
kombu==5.2.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
prompt-toolkit==3.0.37
pylint==2.16.2
pytz==2022.7.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.6
wrapt==1.14.1
